+++ src/stash/src/lib.rs
@@ -9,14 +9,11 @@
 
 //! Durable metadata storage.
 
-use std::collections::btree_map::Entry;
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::{self, Debug};
 use std::marker::PhantomData;
 use std::sync::Arc;
 
-use mz_ore::soft_assert_no_log;
-use mz_proto::{RustType, TryFromProtoError};
 use mz_stash_types::{InternalStashError, StashError};
 use serde::{Deserialize, Serialize};
 use timely::progress::Antichain;
@@ -516,321 +513,5 @@ where
                 })
             })
             .await
-    }
-}
-
-/// TableTransaction emulates some features of a typical SQL transaction over
-/// table for a [`StashCollection`].
-///
-/// It supports:
-/// - uniqueness constraints
-/// - transactional reads and writes (including read-your-writes before commit)
-///
-/// `K` is the primary key type. Multiple entries with the same key are disallowed.
-/// `V` is the an arbitrary value type.
-///
-/// To finalize, add the results of [`TableTransaction::pending()`] to an
-/// [`AppendBatch`].
-#[derive(Debug, PartialEq, Eq)]
-pub struct TableTransaction<K, V> {
-    initial: BTreeMap<K, V>,
-    // The desired state of keys after commit. `None` means the value will be
-    // deleted.
-    pending: BTreeMap<K, Option<V>>,
-    uniqueness_violation: fn(a: &V, b: &V) -> bool,
-}
-
-impl<K, V> TableTransaction<K, V>
-where
-    K: Ord + Eq + Clone,
-    V: Ord + Clone,
-{
-    /// Create a new TableTransaction with initial data. `uniqueness_violation` is a function
-    /// whether there is a uniqueness violation among two values.
-    ///
-    /// Internally the [`Stash`] serializes data as protobuf. All fields in a proto message are
-    /// optional, which makes using them in Rust cumbersome. Generic parameters `KP` and `VP` are
-    /// protobuf types which deserialize to `K` and `V` that a [`TableTransaction`] is generic
-    /// over.
-    pub fn new<KP, VP>(
-        initial: BTreeMap<KP, VP>,
-        uniqueness_violation: fn(a: &V, b: &V) -> bool,
-    ) -> Result<Self, TryFromProtoError>
-    where
-        K: RustType<KP>,
-        V: RustType<VP>,
-    {
-        let initial = initial
-            .into_iter()
-            .map(RustType::from_proto)
-            .collect::<Result<_, _>>()?;
-
-        Ok(Self {
-            initial,
-            pending: BTreeMap::new(),
-            uniqueness_violation,
-        })
-    }
-
-    /// Consumes and returns the pending changes and their diffs. `Diff` is
-    /// guaranteed to be 1 or -1.
-    pub fn pending<KP, VP>(self) -> Vec<(KP, VP, Diff)>
-    where
-        K: RustType<KP>,
-        V: RustType<VP>,
-    {
-        soft_assert_no_log!(self.verify().is_ok());
-        // Pending describes the desired final state for some keys. K,V pairs should be
-        // retracted if they already exist and were deleted or are being updated.
-        self.pending
-            .into_iter()
-            .map(|(k, v)| match self.initial.get(&k) {
-                Some(initial_v) => {
-                    let mut diffs = vec![(k.clone(), initial_v.clone(), -1)];
-                    if let Some(v) = v {
-                        diffs.push((k, v, 1));
-                    }
-                    diffs
-                }
-                None => {
-                    if let Some(v) = v {
-                        vec![(k, v, 1)]
-                    } else {
-                        vec![]
-                    }
-                }
-            })
-            .flatten()
-            .map(|(key, val, diff)| (key.into_proto(), val.into_proto(), diff))
-            .collect()
-    }
-
-    fn verify(&self) -> Result<(), StashError> {
-        // Compare each value to each other value and ensure they are unique.
-        let items = self.items();
-        for (i, vi) in items.values().enumerate() {
-            for (j, vj) in items.values().enumerate() {
-                if i != j && (self.uniqueness_violation)(vi, vj) {
-                    return Err("uniqueness violation".into());
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Iterates over the items viewable in the current transaction in arbitrary
-    /// order and applies `f` on all key, value pairs.
-    pub fn for_values<F: FnMut(&K, &V)>(&self, mut f: F) {
-        let mut seen = BTreeSet::new();
-        for (k, v) in self.pending.iter() {
-            seen.insert(k);
-            // Deleted items don't exist so shouldn't be visited, but still suppress
-            // visiting the key later.
-            if let Some(v) = v {
-                f(k, v);
-            }
-        }
-        for (k, v) in self.initial.iter() {
-            // Add on initial items that don't have updates.
-            if !seen.contains(k) {
-                f(k, v);
-            }
-        }
-    }
-
-    /// Returns the current value of `k`.
-    pub fn get(&self, k: &K) -> Option<&V> {
-        if let Some(v) = self.pending.get(k) {
-            v.as_ref()
-        } else if let Some(v) = self.initial.get(k) {
-            Some(v)
-        } else {
-            None
-        }
-    }
-
-    /// Returns the items viewable in the current transaction.
-    pub fn items(&self) -> BTreeMap<K, V> {
-        let mut items = BTreeMap::new();
-        self.for_values(|k, v| {
-            items.insert(k.clone(), v.clone());
-        });
-        items
-    }
-
-    /// Iterates over the items viewable in the current transaction, and provides a
-    /// map where additional pending items can be inserted, which will be appended
-    /// to current pending items. Does not verify uniqueness.
-    fn for_values_mut<F: FnMut(&mut BTreeMap<K, Option<V>>, &K, &V)>(&mut self, mut f: F) {
-        let mut pending = BTreeMap::new();
-        self.for_values(|k, v| f(&mut pending, k, v));
-        self.pending.extend(pending);
-    }
-
-    /// Inserts a new k,v pair.
-    ///
-    /// Returns an error if the uniqueness check failed or the key already exists.
-    pub fn insert(&mut self, k: K, v: V) -> Result<(), StashError> {
-        let mut violation = None;
-        self.for_values(|for_k, for_v| {
-            if &k == for_k {
-                violation = Some("duplicate key".to_string());
-            }
-            if (self.uniqueness_violation)(for_v, &v) {
-                violation = Some("uniqueness violation".to_string());
-            }
-        });
-        if let Some(violation) = violation {
-            return Err(violation.into());
-        }
-        self.pending.insert(k, Some(v));
-        soft_assert_no_log!(self.verify().is_ok());
-        Ok(())
-    }
-
-    /// Updates k, v pairs. `f` is a function that can return `Some(V)` if the
-    /// value should be updated, otherwise `None`. Returns the number of changed
-    /// entries.
-    ///
-    /// Returns an error if the uniqueness check failed.
-    pub fn update<F: Fn(&K, &V) -> Option<V>>(&mut self, f: F) -> Result<Diff, StashError> {
-        let mut changed = 0;
-        // Keep a copy of pending in case of uniqueness violation.
-        let pending = self.pending.clone();
-        self.for_values_mut(|p, k, v| {
-            if let Some(next) = f(k, v) {
-                changed += 1;
-                p.insert(k.clone(), Some(next));
-            }
-        });
-        // Check for uniqueness violation.
-        if let Err(err) = self.verify() {
-            self.pending = pending;
-            Err(err)
-        } else {
-            Ok(changed)
-        }
-    }
-
-    /// Set the value for a key. Returns the previous entry if the key existed,
-    /// otherwise None.
-    ///
-    /// Returns an error if the uniqueness check failed.
-    ///
-    /// DO NOT call this function in a loop, use [`Self::set_many`] instead.
-    pub fn set(&mut self, k: K, v: Option<V>) -> Result<Option<V>, StashError> {
-        // Save the pending value for the key so we can restore it in case of
-        // uniqueness violation.
-        let restore = self.pending.get(&k).cloned();
-
-        let prev = match self.pending.entry(k.clone()) {
-            // key hasn't been set in this txn yet.
-            Entry::Vacant(e) => {
-                let initial = self.initial.get(&k);
-                if initial != v.as_ref() {
-                    // Provided value and initial value are different. Set key.
-                    e.insert(v);
-                }
-                // Return the initial txn's value of k.
-                initial.cloned()
-            }
-            // key has been set in this txn. Set it and return the previous
-            // pending value.
-            Entry::Occupied(mut e) => e.insert(v),
-        };
-
-        // Check for uniqueness violation.
-        if let Err(err) = self.verify() {
-            // Revert self.pending to the state it was in before calling this
-            // function.
-            match restore {
-                Some(v) => {
-                    self.pending.insert(k, v);
-                }
-                None => {
-                    self.pending.remove(&k);
-                }
-            }
-            Err(err)
-        } else {
-            Ok(prev)
-        }
-    }
-
-    /// Set the values for many keys. Returns the previous entry for each key if the key existed,
-    /// otherwise None.
-    ///
-    /// Returns an error if any uniqueness check failed.
-    pub fn set_many(
-        &mut self,
-        kvs: BTreeMap<K, Option<V>>,
-    ) -> Result<BTreeMap<K, Option<V>>, StashError> {
-        let mut prevs = BTreeMap::new();
-        let mut restores = BTreeMap::new();
-
-        for (k, v) in kvs {
-            // Save the pending value for the key so we can restore it in case of
-            // uniqueness violation.
-            let restore = self.pending.get(&k).cloned();
-            restores.insert(k.clone(), restore);
-
-            let prev = match self.pending.entry(k.clone()) {
-                // key hasn't been set in this txn yet.
-                Entry::Vacant(e) => {
-                    let initial = self.initial.get(&k);
-                    if initial != v.as_ref() {
-                        // Provided value and initial value are different. Set key.
-                        e.insert(v);
-                    }
-                    // Return the initial txn's value of k.
-                    initial.cloned()
-                }
-                // key has been set in this txn. Set it and return the previous
-                // pending value.
-                Entry::Occupied(mut e) => e.insert(v),
-            };
-            prevs.insert(k, prev);
-        }
-
-        // Check for uniqueness violation.
-        if let Err(err) = self.verify() {
-            for (k, restore) in restores {
-                // Revert self.pending to the state it was in before calling this
-                // function.
-                match restore {
-                    Some(v) => {
-                        self.pending.insert(k, v);
-                    }
-                    None => {
-                        self.pending.remove(&k);
-                    }
-                }
-            }
-            Err(err)
-        } else {
-            Ok(prevs)
-        }
-    }
-
-    /// Deletes items for which `f` returns true. Returns the keys and values of
-    /// the deleted entries.
-    pub fn delete<F: Fn(&K, &V) -> bool>(&mut self, f: F) -> Vec<(K, V)> {
-        let mut deleted = Vec::new();
-        self.for_values_mut(|p, k, v| {
-            if f(k, v) {
-                deleted.push((k.clone(), v.clone()));
-                p.insert(k.clone(), None);
-            }
-        });
-        soft_assert_no_log!(self.verify().is_ok());
-        deleted
-    }
-
-    /// Deletes all items.
-    pub fn clear(&mut self) {
-        self.for_values_mut(|p, k, _v| {
-            p.insert(k.clone(), None);
-        });
-        soft_assert_no_log!(self.verify().is_ok());
     }
 }
