+++ src/stash/src/tests.rs
@@ -14,7 +14,7 @@ use std::time::Duration;
 use crate::transaction::MAX_INSERT_ARGUMENTS;
 use crate::{
     AppendBatch, Data, DebugStashFactory, Stash, StashCollection, StashError, StashFactory,
-    TableTransaction, Timestamp, TypedCollection, INSERT_BATCH_SPLIT_SIZE,
+    Timestamp, TypedCollection, INSERT_BATCH_SPLIT_SIZE,
 };
 use futures::Future;
 use mz_ore::assert_contains;
@@ -281,8 +281,6 @@ where
         })
         .await
         .unwrap();
-
-    test_stash_table(&mut stash).await;
 
     stash
 }
@@ -764,290 +762,4 @@ where
     V: Data,
 {
     collection(stash, typed_collection.name()).await
-}
-
-async fn test_stash_table(stash: &mut Stash) {
-    const TABLE: TypedCollection<Vec<u8>, String> = TypedCollection::new("table");
-    fn uniqueness_violation(a: &String, b: &String) -> bool {
-        a == b
-    }
-    let collection = get(&TABLE, stash).await.unwrap();
-
-    async fn commit(
-        stash: &mut Stash,
-        collection: StashCollection<Vec<u8>, String>,
-        pending: Vec<(Vec<u8>, String, i64)>,
-    ) -> Result<(), StashError> {
-        let mut batch = make_batch(&collection, stash).await.unwrap();
-        for (k, v, diff) in pending {
-            collection.append_to_batch(&mut batch, &k, &v, diff);
-        }
-        append(stash, vec![batch]).await.unwrap();
-        Ok(())
-    }
-
-    TABLE
-        .upsert_key(stash, 1i64.to_le_bytes().to_vec(), |_| {
-            Ok::<_, Infallible>("v1".to_string())
-        })
-        .await
-        .unwrap()
-        .unwrap();
-    TABLE
-        .upsert(stash, vec![(2i64.to_le_bytes().to_vec(), "v2".to_string())])
-        .await
-        .unwrap();
-    let mut table =
-        TableTransaction::new(TABLE.peek_one(stash).await.unwrap(), uniqueness_violation).unwrap();
-    assert_eq!(
-        table.items(),
-        BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), "v1".to_string()),
-            (2i64.to_le_bytes().to_vec(), "v2".to_string())
-        ])
-    );
-    assert_eq!(table.delete(|_k, _v| false).len(), 0);
-    assert_eq!(table.delete(|_k, v| v == "v2").len(), 1);
-    assert_eq!(
-        table.items(),
-        BTreeMap::from([(1i64.to_le_bytes().to_vec(), "v1".to_string())])
-    );
-    assert_eq!(table.update(|_k, _v| Some("v3".to_string())).unwrap(), 1);
-
-    // Uniqueness violation.
-    table
-        .insert(3i64.to_le_bytes().to_vec(), "v3".to_string())
-        .unwrap_err();
-
-    table
-        .insert(3i64.to_le_bytes().to_vec(), "v4".to_string())
-        .unwrap();
-    assert_eq!(
-        table.items(),
-        BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), "v3".to_string()),
-            (3i64.to_le_bytes().to_vec(), "v4".to_string()),
-        ])
-    );
-    assert_eq!(
-        table
-            .update(|_k, _v| Some("v1".to_string()))
-            .unwrap_err()
-            .to_string(),
-        "stash error: uniqueness violation"
-    );
-    let pending = table.pending();
-    assert_eq!(
-        pending,
-        vec![
-            (1i64.to_le_bytes().to_vec(), "v1".to_string(), -1),
-            (1i64.to_le_bytes().to_vec(), "v3".to_string(), 1),
-            (2i64.to_le_bytes().to_vec(), "v2".to_string(), -1),
-            (3i64.to_le_bytes().to_vec(), "v4".to_string(), 1),
-        ]
-    );
-    commit(stash, collection, pending).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), "v3".to_string()),
-            (3i64.to_le_bytes().to_vec(), "v4".to_string())
-        ])
-    );
-
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    // Deleting then creating an item that has a uniqueness violation should work.
-    assert_eq!(table.delete(|k, _v| k == &1i64.to_le_bytes()).len(), 1);
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v3".to_string())
-        .unwrap();
-    // Uniqueness violation in value.
-    table
-        .insert(5i64.to_le_bytes().to_vec(), "v3".to_string())
-        .unwrap_err();
-    // Key already exists, expect error.
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v5".to_string())
-        .unwrap_err();
-    assert_eq!(table.delete(|k, _v| k == &1i64.to_le_bytes()).len(), 1);
-    // Both the inserts work now because the key and uniqueness violation are gone.
-    table
-        .insert(5i64.to_le_bytes().to_vec(), "v3".to_string())
-        .unwrap();
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v5".to_string())
-        .unwrap();
-    let pending = table.pending();
-    assert_eq!(
-        pending,
-        vec![
-            (1i64.to_le_bytes().to_vec(), "v3".to_string(), -1),
-            (1i64.to_le_bytes().to_vec(), "v5".to_string(), 1),
-            (5i64.to_le_bytes().to_vec(), "v3".to_string(), 1),
-        ]
-    );
-    commit(stash, collection, pending).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), "v5".to_string()),
-            (3i64.to_le_bytes().to_vec(), "v4".to_string()),
-            (5i64.to_le_bytes().to_vec(), "v3".to_string()),
-        ])
-    );
-
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    assert_eq!(table.delete(|_k, _v| true).len(), 3);
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v1".to_string())
-        .unwrap();
-
-    commit(stash, collection, table.pending()).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([(1i64.to_le_bytes().to_vec(), "v1".to_string()),])
-    );
-
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    assert_eq!(table.delete(|_k, _v| true).len(), 1);
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v2".to_string())
-        .unwrap();
-    commit(stash, collection, table.pending()).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([(1i64.to_le_bytes().to_vec(), "v2".to_string()),])
-    );
-
-    // Verify we don't try to delete v3 or v4 during commit.
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    assert_eq!(table.delete(|_k, _v| true).len(), 1);
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v3".to_string())
-        .unwrap();
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v4".to_string())
-        .unwrap_err();
-    assert_eq!(table.delete(|_k, _v| true).len(), 1);
-    table
-        .insert(1i64.to_le_bytes().to_vec(), "v5".to_string())
-        .unwrap();
-    commit(stash, collection, table.pending()).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items.into_iter().collect::<Vec<_>>(),
-        vec![(1i64.to_le_bytes().to_vec(), "v5".to_string())]
-    );
-
-    // Test `set`.
-    let items = TABLE.peek_one(stash).await.unwrap();
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    // Uniqueness violation.
-    table
-        .set(2i64.to_le_bytes().to_vec(), Some("v5".to_string()))
-        .unwrap_err();
-    table
-        .set(3i64.to_le_bytes().to_vec(), Some("v6".to_string()))
-        .unwrap();
-    table.set(2i64.to_le_bytes().to_vec(), None).unwrap();
-    table.set(1i64.to_le_bytes().to_vec(), None).unwrap();
-    let pending = table.pending();
-    assert_eq!(
-        pending,
-        vec![
-            (1i64.to_le_bytes().to_vec(), "v5".to_string(), -1),
-            (3i64.to_le_bytes().to_vec(), "v6".to_string(), 1),
-        ]
-    );
-    commit(stash, collection, pending).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([(3i64.to_le_bytes().to_vec(), "v6".to_string())])
-    );
-
-    // Duplicate `set`.
-    let items = TABLE.peek_one(stash).await.unwrap();
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    table
-        .set(3i64.to_le_bytes().to_vec(), Some("v6".to_string()))
-        .unwrap();
-    let pending = table.pending::<Vec<u8>, String>();
-    assert!(pending.is_empty());
-
-    // Test `set_many`.
-    let items = TABLE.peek_one(stash).await.unwrap();
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    // Uniqueness violation.
-    table
-        .set_many(BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), Some("v6".to_string())),
-            (42i64.to_le_bytes().to_vec(), Some("v1".to_string())),
-        ]))
-        .unwrap_err();
-    table
-        .set_many(BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), Some("v6".to_string())),
-            (3i64.to_le_bytes().to_vec(), Some("v1".to_string())),
-        ]))
-        .unwrap();
-    table
-        .set_many(BTreeMap::from([
-            (42i64.to_le_bytes().to_vec(), Some("v7".to_string())),
-            (3i64.to_le_bytes().to_vec(), None),
-        ]))
-        .unwrap();
-    let pending = table.pending();
-    assert_eq!(
-        pending,
-        vec![
-            (1i64.to_le_bytes().to_vec(), "v6".to_string(), 1),
-            (3i64.to_le_bytes().to_vec(), "v6".to_string(), -1),
-            (42i64.to_le_bytes().to_vec(), "v7".to_string(), 1),
-        ]
-    );
-    commit(stash, collection, pending).await.unwrap();
-    let items = TABLE.peek_one(stash).await.unwrap();
-    assert_eq!(
-        items,
-        BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), "v6".to_string()),
-            (42i64.to_le_bytes().to_vec(), "v7".to_string())
-        ])
-    );
-
-    // Duplicate `set_many`.
-    let items = TABLE.peek_one(stash).await.unwrap();
-    let mut table = TableTransaction::new(items, uniqueness_violation).unwrap();
-    table
-        .set_many(BTreeMap::from([
-            (1i64.to_le_bytes().to_vec(), Some("v6".to_string())),
-            (42i64.to_le_bytes().to_vec(), Some("v7".to_string())),
-        ]))
-        .unwrap();
-    let pending = table.pending::<Vec<u8>, String>();
-    assert!(pending.is_empty());
-}
-
-#[mz_ore::test]
-fn test_table() {
-    fn uniqueness_violation(a: &String, b: &String) -> bool {
-        a == b
-    }
-    let mut table = TableTransaction::new(
-        BTreeMap::from([(1i64.to_le_bytes().to_vec(), "a".to_string())]),
-        uniqueness_violation,
-    )
-    .unwrap();
-
-    table
-        .insert(2i64.to_le_bytes().to_vec(), "b".to_string())
-        .unwrap();
-    table
-        .insert(3i64.to_le_bytes().to_vec(), "c".to_string())
-        .unwrap();
 }
